backend/requirements.txt
fastapi==0.133.1
uvicorn==0.41.0
sqlalchemy==2.0.47
psycopg2-binary==2.9.11
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.5
openai==2.24.0
langchain-community==0.4.1
sentence-transformers==5.2.3
faiss-cpu==1.13.2
scikit-learn==1.8.0
pandas==3.0.1
numpy==2.4.2
matplotlib
seaborn
python-multipart==0.0.22
torch
torchvision
torchaudio
